Home component is now retrieving access token
