+++ src/App/App-reducers.js
@@ -24,3 +24,4 @@ export function apiSuccess(state={}, action) {
       return state
   }
 }
+
