+++ src/Home/Home-actions.js
@@ -1,0 +1,6 @@
+export const saveAccessTokenAction = (token) => {
+	return {
+		type: 'ACCESS_TOKEN',
+		token
+	}
+}
