+++ src/Home/Home-reducers.js
@@ -1,0 +1,8 @@
+export function accessToken(state='', action) {
+	switch(action.type) {
+		case 'ACCESS_TOKEN':
+			return action
+		default:
+			return state
+	}
+}
